Improve management of bitstreams used for testing
